# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Bank Cards management system
  description: API Description for Bank Cards management system
  contact:
    name: Alex VM
    email: my@email.com
  version: 0.0.1
servers:
- url: http://localhost:8080
  description: Generated server url
paths:
...
  /api/v1/users/admin/update/{userId}:
    patch:
      tags:
      - Users
      summary: Изменение Админом данных Юзера по его ID
      operationId: updateUser
      parameters:
      - name: userId
        in: path
        required: true
        schema:
          type: integer
          format: int64
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserUpdateDto"
        required: true
      responses:
        "200":
          description: User updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserRespDto"
        "400":
          $ref: "#/components/responses/Validation"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
        "401":
          description: Authentication required
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponseSchema"
        "403":
          description: Access denied
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponseSchema"
      security:
      - JWT Bearer: []
...
  /api/v1/users/me:
    get:
      tags:
      - Users
      summary: Получение данных текущего Юзера (для тестирования)
      operationId: authenticatedUser
      responses:
        "200":
          description: User found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserRespDto"
        "401":
          description: Authentication required
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponseSchema"
        "403":
          description: Access denied
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponseSchema"
      security:
      - JWT Bearer: []
  /api/v1/users/admin/{userId}:
    get:
      tags:
      - Users
      summary: Получение Админом данных Юзера по его ID
      operationId: getUserById
      parameters:
      - name: userId
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: User found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserRespDto"
        "404":
          $ref: "#/components/responses/NotFound"
        "401":
          description: Authentication required
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponseSchema"
        "403":
          description: Access denied
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponseSchema"
      security:
      - JWT Bearer: []
  /api/v1/users/admin/list:
    get:
      tags:
      - Users
      summary: Получение Админом списка Юзеров (с фильтрацией и пагинацией)
      description: При отсутствии параметров фильтрации возвращается список ВСЕХ Юзеров.
      operationId: getUsers
      parameters:
      - name: dto
        in: query
        required: true
        schema:
          $ref: "#/components/schemas/UserFilterDto"
      - name: page
        in: query
        description: Zero-based page index (0..N)
        required: false
        schema:
          type: integer
          default: 0
          minimum: 0
      - name: size
        in: query
        description: The size of the page to be returned
        required: false
        schema:
          type: integer
          default: 10
          minimum: 1
      - name: sort
        in: query
        description: "Sorting criteria in the format: property,(asc|desc). Default\
          \ sort order is ascending. Multiple sort criteria are supported."
        required: false
        schema:
          type: array
          default:
          - "id,ASC"
          items:
            type: string
      responses:
        "200":
          description: Page of users
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserPageResponse"
        "401":
          description: Authentication required
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponseSchema"
        "403":
          description: Access denied
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponseSchema"
      security:
      - JWT Bearer: []
...
components:
  schemas:
...
    ErrorResponseSchema:
      type: object
      description: Unified error response
      properties:
        type:
          type: string
          description: "The string containing a URI reference [URI] that identifies\
            \ the problem type."
          example: about:blank
        title:
          type: string
          example: VALIDATION_ERROR
        status:
          type: integer
          format: int32
          description: The number indicating the HTTP status code
          example: 400
        detail:
          type: string
          example: Validation failed
        instance:
          type: string
          description: The string containing a URI reference that identifies the specific
            occurrence of the problem
          example: /users/admin
        errors:
          type: object
          additionalProperties:
            type: string
          description: Validation errors (only for VALIDATION_ERROR)
          example:
            email: must be a valid email
...
    UserRespDto:
      type: object
      properties:
        id:
          type: integer
          format: int64
        email:
          type: string
        role:
          type: string
          enum:
          - USER
          - ADMIN
        firstName:
          type: string
        lastName:
          type: string
...
    UserUpdateDto:
      type: object
      description: Частичное обновление пользователя
      properties:
        email:
          type: string
          format: email
          description: email
          example: email@email.com
        firstName:
          type: string
          description: First Name to print it on card
          example: TOM
          maxLength: 20
          minLength: 2
          pattern: "[A-Za-z]+"
        lastName:
          type: string
          description: First Name to print it on card
          example: SMITH
          maxLength: 20
          minLength: 2
          pattern: "[A-Za-z]+"
        role:
          type: string
          description: User role
          enum:
          - USER
          - ADMIN
          example: USER/ADMIN
...
    UserFilterDto:
      type: object
      properties:
        firstName:
          type: string
          pattern: "[A-Za-z]+"
        lastName:
          type: string
          pattern: "[A-Za-z]+"
        email:
          type: string
          format: email
        role:
          type: string
          enum:
          - USER
          - ADMIN
    PageMeta:
      type: object
      description: Pagination metadata
      properties:
        size:
          type: integer
          format: int32
          example: 10
        number:
          type: integer
          format: int32
          example: 0
        totalElements:
          type: integer
          format: int64
          example: 2
        totalPages:
          type: integer
          format: int32
          example: 1
    UserPageResponse:
      type: object
      properties:
        content:
          type: array
          description: Page content
          items:
            $ref: "#/components/schemas/UserRespDto"
        page:
          $ref: "#/components/schemas/PageMeta"
          description: Pagination info
...
  responses:
    NotFound:
      description: Entity not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponseSchema"
...
    Conflict:
      description: Conflict
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponseSchema"
    Validation:
      description: Validation error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponseSchema"
  securitySchemes:
    JWT Bearer:
      type: http
      description: JWT in header
      in: header
      scheme: bearer
      bearerFormat: JWT